requirements-resume-studio.txt
streamlit>=1.44.0
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pypdf>=5.4.0
python-docx>=1.1.2
docx2txt>=0.8
reportlab>=4.0.0
